requirements.txt
requests==2.22.0
python_telegram_bot==12.5.1
